Rename component and page files for property details page
