+++ resources/js/Components/PropertyCarousel.jsx
